WIP: Removed unused methods from the SextantNavigationServiceBase Splited the NavigationElement to another file
